+++ dictionary/src/main/java/neoflex/dictionary/service/DictionaryService.java
@@ -20,12 +20,26 @@ public class DictionaryService {
         return dictionaryRepository.findAll();
     }
 
-    public Dictionary saveDictionary(Dictionary dictionary){
-        return dictionaryRepository.save(dictionary);
+    public String saveDictionary(Dictionary dictionary){
+        if (dictionary == null){
+            dictionary = Dictionary.defaultValue;
+        }
+        dictionaryRepository.save(dictionary);
+        return "Dictionary saved";
     }
 
-    public void deleteDictionary(Dictionary dictionary){
+    public String deleteDictionary(Dictionary dictionary){
+        if (dictionary == null){
+            dictionary = findFirst();
+            if (dictionary == null){
+                return "All dictionaries already deleted";
+            }
+        }
+        if (findDictionary(dictionary) == null){
+            return "Dictionary does not exist";
+        }
         dictionaryRepository.delete(dictionary);
+        return "Dictionary deleted";
     }
 
     public Dictionary findFirst()
